fix(doom/am_map): compute the automap clear size in size_t
AM_clearFB does memset(fb, color, f_w*f_h). f_w and f_h are int, so the size
is computed in 32-bit and only widened to memset's size_t afterwards — it can
wrap before the conversion. CodeQL flags this as
cpp/integer-multiplication-cast-to-long (CWE-190).

Cast the first operand to size_t so the whole product is 64-bit. No behavioural
change at real automap dimensions; the arithmetic is just correct if they grow.

+++ kernel/doom_src/am_map.c
@@ -833,7 +833,7 @@ void AM_Ticker (void)
 //
 void AM_clearFB(int color)
 {
-    memset(fb, color, f_w*f_h);
+    memset(fb, color, (size_t)f_w*f_h);   // 64-bit multiply: f_w*f_h in int can overflow before widening to memset's size_t
 }
